Rename hexagon_anal
