+++ librz/analysis/meson.build
@@ -96,7 +96,7 @@ rz_analysis_sources = [
   'p/analysis_xtensa.c',
   'p/analysis_z80.c',
   #'arch/gb/meta_gb_cmt.c',
-  'arch/hexagon/hexagon_anal.c',
+  'arch/hexagon/hexagon_analysis.c',
   '../asm/arch/amd29k/amd29k.c',
   '../asm/arch/arm/winedbg/be_arm.c',
   '../asm/arch/cr16/cr16_disas.c',
